CE-6909: move class MaterialInfo into its own file
